Move reporters/colorizer.h to colorizers/colorizer.h
