+++ bandit/colorizers/colorizer.h
